pkp/pkp-lib#9887 Add new discovery pattern
