+++ classes/invitation/core/Invitation.php
@@ -1,7 +1,7 @@
 <?php
 
 /**
- * @file classes/invitation/invitations/Invitation.php
+ * @file classes/invitation/core/Invitation.php
  *
  * Copyright (c) 2023 Simon Fraser University
  * Copyright (c) 2023 John Willinsky
@@ -12,17 +12,17 @@
  * @brief Abstract class for all Invitations
  */
 
-namespace PKP\invitation\invitations;
+namespace PKP\invitation\core;
 
 use Carbon\Carbon;
 use Exception;
 use Illuminate\Support\Facades\Mail;
 use PKP\config\Config;
-use PKP\invitation\invitations\enums\InvitationAction;
-use PKP\invitation\invitations\enums\InvitationStatus;
-use PKP\invitation\invitations\PKPInvitationActionRedirectController;
-use PKP\invitation\invitations\traits\HasMailable;
-use PKP\invitation\invitations\traits\ShouldValidate;
+use PKP\invitation\core\enums\InvitationAction;
+use PKP\invitation\core\enums\InvitationStatus;
+use PKP\invitation\core\PKPInvitationActionRedirectController;
+use PKP\invitation\core\traits\HasMailable;
+use PKP\invitation\core\traits\ShouldValidate;
 use PKP\invitation\models\InvitationModel;
 use PKP\pages\invitation\PKPInvitationHandler;
 use PKP\security\Validation;
