+++ classes/invitation/core/PKPReceiveInvitationController.php
@@ -1,18 +1,18 @@
 <?php
 
 /**
- * @file classes/invitation/invitations/PKPCreateInvitationController.php
+ * @file classes/invitation/core/PKPReceiveInvitationController.php
  *
  * Copyright (c) 2023 Simon Fraser University
  * Copyright (c) 2023 John Willinsky
  * Distributed under the GNU GPL v3. For full terms see the file docs/COPYING.
  *
- * @class PKPCreateInvitationController
+ * @class PKPReceiveInvitationController
  *
  * @brief Interface for all Invitation API Handlers
  */
 
-namespace PKP\invitation\invitations;
+namespace PKP\invitation\core;
 
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
@@ -20,11 +20,11 @@ use Illuminate\Routing\Controller;
 use PKP\core\PKPBaseController;
 use PKP\core\PKPRequest;
 
-abstract class PKPCreateInvitationController extends Controller
+abstract class PKPReceiveInvitationController extends Controller
 {
     abstract public function authorize(PKPBaseController $controller, PKPRequest $request, array &$args, array $roleAssignments): bool;
-    abstract public function add(Request $illuminateRequest): JsonResponse;
-    abstract public function populate(Request $illuminateRequest): JsonResponse;
-    abstract public function dispatch(Request $illuminateRequest): JsonResponse;
-    abstract public function get(Request $illuminateRequest): JsonResponse;
+    abstract public function receive(Request $illuminateRequest): JsonResponse;
+    abstract public function refine(Request $illuminateRequest): JsonResponse;
+    abstract public function finalize(Request $illuminateRequest): JsonResponse;
+    abstract public function decline(Request $illuminateRequest): JsonResponse;
 }
