+++ classes/invitation/core/enums/InvitationAction.php
@@ -1,7 +1,7 @@
 <?php
 
 /**
- * @file invitation/invitations/enums/InvitationAction.php
+ * @file invitation/core/enums/InvitationAction.php
  *
  * Copyright (c) 2023 Simon Fraser University
  * Copyright (c) 2023 John Willinsky
@@ -12,7 +12,7 @@
  * @brief Enumeration for invitation actions
  */
 
-namespace PKP\invitation\invitations\enums;
+namespace PKP\invitation\core\enums;
 
 enum InvitationAction: string
 {
