+++ classes/invitation/invitations/handlers/ChangeProfileEmailInviteRedirectController.php
@@ -1,33 +1,31 @@
 <?php
 
 /**
- * @file invitations/handlers/RegistrationAccessInviteRedirectController.php
+ * @file classes/invitation/invitations/handlers/ChangeProfileEmailInviteRedirectController.php
  *
  * Copyright (c) 2023 Simon Fraser University
  * Copyright (c) 2023 John Willinsky
  * Distributed under the GNU GPL v3. For full terms see the file docs/COPYING.
  *
- * @class RegistrationAccessInviteRedirectController
- *
- * @ingroup invitations\handlers
+ * @class ChangeProfileEmailInviteRedirectController
  *
  * @brief Change Profile Email invitation
  */
 
-namespace PKP\invitations\handlers;
+namespace PKP\invitation\invitations\handlers;
 
 use APP\core\Request;
-use APP\facades\Repo;
-use Exception;
+use APP\notification\NotificationManager;
 use PKP\core\PKPApplication;
-use PKP\invitation\invitations\enums\InvitationAction;
-use PKP\invitation\invitations\enums\InvitationStatus;
-use PKP\invitation\invitations\PKPInvitationActionRedirectController;
-use PKP\invitations\RegistrationAccessInvite;
+use APP\facades\Repo;
+use PKP\invitation\core\enums\InvitationAction;
+use PKP\invitation\core\enums\InvitationStatus;
+use PKP\invitation\core\PKPInvitationActionRedirectController;
+use PKP\invitation\invitations\ChangeProfileEmailInvite;
 
-class RegistrationAccessInviteRedirectController extends PKPInvitationActionRedirectController
+class ChangeProfileEmailInviteRedirectController extends PKPInvitationActionRedirectController
 {
-    public function getInvitation(): RegistrationAccessInvite
+    public function getInvitation(): ChangeProfileEmailInvite
     {
         return $this->invitation;
     }
@@ -37,38 +35,22 @@ class RegistrationAccessInviteRedirectController extends PKPInvitationActionRedi
         if ($this->invitation->getStatus() !== InvitationStatus::ACCEPTED) {
             $request->getDispatcher()->handle404();
         }
-        
-        $user = Repo::user()->get($this->invitation->invitationModel->userId, true);
-        
-        if (!$user) {
-            throw new Exception();
-        }
+
+        $user = Repo::user()->get($this->invitation->invitationModel->userId);
+
+        $notificationManager = new NotificationManager();
+        $notificationManager->createTrivialNotification($user->getId());
 
         $url = PKPApplication::get()->getDispatcher()->url(
             PKPApplication::get()->getRequest(),
             PKPApplication::ROUTE_PAGE,
             null,
             'user',
-            'activateUser',
+            'profile',
             [
-                $user->getUsername(),
+                'contact'
             ]
         );
-
-        if (isset($this->invitationModel->contextId)) {
-            $context = $request->getContext();
-
-            $url = PKPApplication::get()->getDispatcher()->url(
-                PKPApplication::get()->getRequest(),
-                PKPApplication::ROUTE_PAGE,
-                $context->getData('urlPath'),
-                'user',
-                'activateUser',
-                [
-                    $user->getUsername(),
-                ]
-            );
-        }
 
         $request->redirectUrl($url);
     }
@@ -79,22 +61,25 @@ class RegistrationAccessInviteRedirectController extends PKPInvitationActionRedi
             $request->getDispatcher()->handle404();
         }
 
-        $context = $request->getContext();
+        $user = Repo::user()->get($this->invitation->invitationModel->userId);
+
+        $notificationManager = new NotificationManager();
+        $notificationManager->createTrivialNotification($user->getId());
 
         $url = PKPApplication::get()->getDispatcher()->url(
             PKPApplication::get()->getRequest(),
             PKPApplication::ROUTE_PAGE,
-            $context->getData('urlPath'),
-            'user',
-            'login',
             null,
+            'user',
+            'profile',
             [
+                'contact'
             ]
         );
 
         $request->redirectUrl($url);
     }
-    
+
     public function preRedirectActions(InvitationAction $action) 
     {
         if ($action == InvitationAction::ACCEPT) {
