+++ classes/observers/listeners/ValidateRegisteredEmail.php
@@ -20,7 +20,7 @@ use APP\facades\Repo;
 use Illuminate\Events\Dispatcher;
 use Illuminate\Support\Facades\Mail;
 use PKP\config\Config;
-use PKP\invitations\RegistrationAccessInvite;
+use PKP\invitation\invitations\RegistrationAccessInvite;
 use PKP\mail\mailables\ValidateEmailContext as ContextMailable;
 use PKP\mail\mailables\ValidateEmailSite as SiteMailable;
 use PKP\observers\events\UserRegisteredContext;
